Successfully seperated the applications into a infra and handlers folder. This done for ease of navigating through the codebase.
