+++ handlers/package.json
@@ -26,6 +26,8 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@aws-sdk/client-dynamodb": "^3.484.0",
+		"@aws-sdk/lib-dynamodb": "^3.484.0",
 		"@types/aws-lambda": "^8.10.130",
 		"esbuild": "^0.19.9",
 		"uuid": "^9.0.1"
