+++ infra/lib/storage_stack_service.ts
