+++ infra/package.json
@@ -10,11 +10,9 @@
 		"test": "jest",
 		"cdk": "cdk",
 		"integration-test": "npx playwright test --workers=1",
-		"deploy": "cd resources && npm run build && cd dist && zip -r ../dist.zip * && cd ../.. && cdk synth && cdk deploy --all --require-approval='never'"
+		"deploy": "cd ../handlers && npm run build && cd dist && zip -r ../dist.zip * && cd ../../infra && cdk synth --app 'npx ts-node bin/bolleje-temporarysecret-cdk.ts' && cdk deploy --all --require-approval='never'"
 	},
 	"devDependencies": {
-		"@aws-sdk/client-dynamodb": "^3.473.0",
-		"@aws-sdk/lib-dynamodb": "^3.473.0",
 		"@types/jest": "^29.5.8",
 		"@types/node": "20.9.1",
 		"aws-cdk": "2.114.1",
