+++ infra/types/enums.ts
@@ -1,0 +1,6 @@
+export enum eMethods {
+    GET_SHE_SECRET = "getSHE",
+    POST_SHE_SECRET = "addSHE",
+    GET_E2E_SECRET = "getE2E",
+    POST_E2E_SECRET = "addE2E",
+}
